refactor(Arena): Put the each arenas into its own namespace
