+++ Makefile
@@ -37,6 +37,7 @@ MAIN_SRC:= \
 
 ARENA_SRC:= \
 	Arena/Arena.cpp \
+	Arena/Container.cpp \
 	Arena/main.cpp
 
 URI_SRC:= \
