+++ src/Arena/Container.cpp
@@ -1,0 +1,5 @@
+#include "Container.hpp"
+
+Arena permanent::arena(1024);
+Arena temporary::arena(1024);
+
